Code Cleanup: Spacing, Moving Code Around, Etc.
